+++ lib/db/index.d.ts
@@ -9,3 +9,4 @@ export * from './gql-pubsub-sequelize-engine';
 export * from './db-to-gql';
 export * from './create-auth-where-clause';
 export * from './convert-to-sequelize-date-filters';
+export * from './db.convict';