+++ src/db/init-sequelize.ts
@@ -1,5 +1,6 @@
-import { Sequelize, Model, ModelCtor } from 'sequelize-typescript';
+import { Sequelize, Model, ModelCtor, SequelizeOptions } from 'sequelize-typescript';
 import { Oid } from '@rumbleship/oid';
+import { RumbleshipDatabaseOptions } from './db.convict';
 
 let theSequelizeInstance: Sequelize | null;
 /**
@@ -35,8 +36,8 @@ export function getOidFor(model: Model) {
 }
 /**
  * initializes sequelize for the app and sets up a global sequelize
- * @param config An object with the shape: { db: { database, username, password, dialect, host, port, pool, socketPath, define }}
- * This is modified and passed to new Sequelize()
+ * @param cfg A Sequelize Configuration object, passed directly, or ((DEPRECATED)) nested under a key `{db: cfg}`
+ * { database, username, password, dialect, host, port, pool, define, dialectOptions: { socketPath, maxPreparedStatements }}
  * @param loggingFun The logging function to pass in. Required
  * @param dbModels An array of sequelize-typescript models for this application
  * @param opt options...
@@ -46,7 +47,7 @@ export function getOidFor(model: Model) {
  * pubSub is the PubSubEngine to use to publish model changes
  */
 export async function initSequelize(
-  config: any,
+  cfg: RumbleshipDatabaseOptions | { db: RumbleshipDatabaseOptions },
   loggingFun: (msg: string) => any,
   dbModels: DbModelAndOidScope[],
   opt?: {
@@ -54,36 +55,29 @@ export async function initSequelize(
     dbSuffix?: string;
   }
 ): Promise<Sequelize> {
-  const force = opt ? opt.force : false;
-  const dbSuffix = opt ? (opt.dbSuffix ? opt.dbSuffix : '') : '';
-  const {
-    db: { database, username, password, dialect, host, port, pool, socketPath, define }
-  } = config;
-  const db = database + dbSuffix;
+  // Forward/backward compatibility with old format of passing entire app config in.
+  const config: RumbleshipDatabaseOptions = Reflect.get(cfg, 'db') ?? cfg;
+  const force = opt?.force ?? false;
+  const dbSuffix = opt?.dbSuffix ?? '';
+  const { database, username, password, dialect, host, port, pool, define } = config;
+  const { socketPath, ...dialectOptions } = config.dialectOptions;
+  const db = database + '_' + dbSuffix;
 
-  let options;
+  const options: SequelizeOptions = {
+    database: db,
+    username,
+    password,
+    dialect,
+    pool,
+    logging: loggingFun,
+    define,
+    dialectOptions
+  };
   if (socketPath) {
-    options = {
-      database: db,
-      username,
-      password,
-      dialect,
-      pool,
-      logging: loggingFun,
-      define,
-      dialectOptions: { socketPath, maxPreparedStatements: 10 }
-    };
+    Reflect.set(options.dialectOptions!, 'socketPath', socketPath);
   } else {
-    options = {
-      database: db,
-      username,
-      password,
-      dialect,
-      logging: loggingFun,
-      define,
-      host,
-      port
-    };
+    options.host = host;
+    options.port = port;
   }
 
   if (['test', 'development'].includes(process.env.NODE_ENV as string) && dbSuffix.length) {