+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+### Removed
+### Changed
+  * `ctx.release()` now async; collects and sends interesting mysql variables to trace on close
+### Fixed
+### Deprecated
+### Security
+
+
 ## [7.0.5] -- 2020-05-15
 
 ### Fixed